requirements.txt
mega.py
lbry-libtorrent